Core updates
Remove dependency of the folder  when calling the  method.
This will allow for a more dynamic folder structure.

Add an  parameter to be passed in to the  method to let the developer
define the root path of the JS files they wish to dynamically load.
Update documentation.
Edit travis to run tests when a new version is pushed and tags